change to original type for config

+++ zlgcan/examples/utils/mod.rs
@@ -9,14 +9,14 @@ pub fn init_device() -> Result<ZCanDriver, CanError> {
     let mut builder = DeviceBuilder::new();
 
     let mut ch1_cfg = ChannelConfig::new(500_000);
-    ch1_cfg.add_other(CHANNEL_MODE, Box::new(ZCanChlMode::Normal as u8))
-        .add_other(CHANNEL_TYPE, Box::new(ZCanChlType::CAN as u8));
+    ch1_cfg.add_other(CHANNEL_MODE, Box::new(ZCanChlMode::Normal))
+        .add_other(CHANNEL_TYPE, Box::new(ZCanChlType::CAN));
 
     let mut ch2_cfg = ChannelConfig::new(500_000);
-    ch2_cfg.add_other(CHANNEL_MODE, Box::new(ZCanChlMode::Normal as u8))
-        .add_other(CHANNEL_TYPE, Box::new(ZCanChlType::CAN as u8));
+    ch2_cfg.add_other(CHANNEL_MODE, Box::new(ZCanChlMode::Normal))
+        .add_other(CHANNEL_TYPE, Box::new(ZCanChlType::CAN));
 
-    builder.add_other(DEVICE_TYPE, Box::new(dev_type as u32))
+    builder.add_other(DEVICE_TYPE, Box::new(dev_type))
         .add_other(DEVICE_INDEX, Box::new(0))
         .add_config(0.to_string(), ch1_cfg)
         .add_config(1.to_string(), ch2_cfg);
